+++ lib/oxml.rb
@@ -37,6 +37,4 @@ module OXML
 
     StringIO.new(xml)
   end
-
-  module_function :optimize_xml_input
 end